kafka_slack_consumer/requirements.txt
slack_sdk==3.27.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
